ts bot db connection and types working

+++ bot/src/bot.ts
@@ -1,3 +1,8 @@
+import { Message, TextChannel } from "discord.js";
+import mongoose from "mongoose";
+import verificationCodes from "./models/verificationCodesModel";
+import owners from "./models/ownersModel";
+import { verifCodesSchema } from "./utils/interface";
 import dotenv from "dotenv";
 dotenv.config();
 
@@ -7,15 +12,21 @@ const client = new Client({
 });
 const PREFIX = ".";
 
-import verificationCodes from "./models/verificationCodesModel";
-import owners from "./models/ownersModel";
-import { verifCodesSchema } from "./utils/interface";
+mongoose
+  .connect(process.env.DB_CONNECT!, {
+    useNewUrlParser: true,
+    useUnifiedTopology: true,
+    useFindAndModify: false,
+  })
+  .catch((err) => {
+    console.log(err);
+  });
 
 client.on("ready", () => {
   console.log(`${client.user.tag} has logged in.`);
 });
 
-client.on("message", async (message:any) => {
+client.on("message", async (message:Message) => {
   if (message.author.bot) return;
   console.log(`[${message.author.tag}]: ${message.content}`);
   if (message.content.startsWith(PREFIX)) {
@@ -32,23 +43,29 @@ client.on("message", async (message:any) => {
     Math.random().toString(36).substring(7);
 
     if (cmd_name === "register") {
-      const isDuplicates = await verificationCodes.find({
-        discordTag: message.author.tag,
-      });
 
+      if (isDM) return message.author.send(
+        "'.register' only works in servers."
+      );
       if (!isDM) {
         message.delete();
         message.author.send(
           "DM me '.register `YourEmail@example.com`' to start/continue the verification process."
         );
 
+        const isDuplicates = await verificationCodes.find({
+          discordTag: message.author.tag,
+        });
+
+        console.log(isDuplicates);
+
         let verif = new verificationCodes({
           discordId: message.author.id,
           discordTag: message.author.tag,
           email: "",
           avatar: message.author.displayAvatarURL(),
-          serverId: message.guild.id,
-          serverName: message.guild.name,
+          serverId: message.guild!.id,
+          serverName: message.guild!.name,
           verificationCode: verificationCode,
           time: new Date().toUTCString(),
         });
@@ -56,8 +73,8 @@ client.on("message", async (message:any) => {
       }
     } else if (cmd_name === "clear") {
       const numArg = parseInt(args[0]);
-      if (numArg > 0) return message.channel.bulkDelete(numArg);
-      return message.channel.bulkDelete(5);
+      if (numArg > 0) return (message.channel as TextChannel).bulkDelete(numArg);
+      return (message.channel as TextChannel).bulkDelete(5);
     }
   }
 });
